Renamed AccountDto, and made UpdateAccountDto extend EntityDto
