+++ src/main/kotlin/pt/up/fe/ni/website/backend/controller/AccountController.kt
@@ -8,9 +8,9 @@ import org.springframework.web.bind.annotation.PutMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import pt.up.fe.ni.website.backend.dto.account.UpdateAccountDto
 import pt.up.fe.ni.website.backend.dto.auth.ChangePasswordDto
-import pt.up.fe.ni.website.backend.dto.entity.AccountDto
+import pt.up.fe.ni.website.backend.dto.entity.account.CreateAccountDto
+import pt.up.fe.ni.website.backend.dto.entity.account.UpdateAccountDto
 import pt.up.fe.ni.website.backend.service.AccountService
 
 @RestController
@@ -23,7 +23,7 @@ class AccountController(private val service: AccountService) {
     fun getAccountById(@PathVariable id: Long) = service.getAccountById(id)
 
     @PostMapping("/new")
-    fun createAccount(@RequestBody dto: AccountDto) = service.createAccount(dto)
+    fun createAccount(@RequestBody dto: CreateAccountDto) = service.createAccount(dto)
 
     @PostMapping("/changePassword/{id}")
     fun changePassword(@PathVariable id: Long, @RequestBody dto: ChangePasswordDto): Map<String, String> {
@@ -38,5 +38,8 @@ class AccountController(private val service: AccountService) {
     ) = service.updateAccountById(id, dto)
 
     @DeleteMapping("/{id}")
-    fun deleteAccountById(@PathVariable id: Long) = service.deleteAccountById(id)
+    fun deleteAccountById(@PathVariable id: Long): Map<String, String> {
+        service.deleteAccountById(id)
+        return emptyMap()
+    }
 }
