+++ src/main/kotlin/pt/up/fe/ni/website/backend/dto/entity/account/CreateAccountDto.kt
@@ -1,10 +1,13 @@
-package pt.up.fe.ni.website.backend.dto.account
+package pt.up.fe.ni.website.backend.dto.entity.account
 
 import java.util.Date
 import pt.up.fe.ni.website.backend.dto.entity.CustomWebsiteDto
+import pt.up.fe.ni.website.backend.dto.entity.EntityDto
+import pt.up.fe.ni.website.backend.model.Account
 
-class UpdateAccountDto(
+class CreateAccountDto(
     val email: String,
+    val password: String,
     val name: String,
     val bio: String?,
     val birthDate: Date?,
@@ -12,4 +15,4 @@ class UpdateAccountDto(
     val linkedin: String?,
     val github: String?,
     val websites: List<CustomWebsiteDto>?
-)
+) : EntityDto<Account>()
