+++ src/main/kotlin/pt/up/fe/ni/website/backend/dto/entity/account/UpdateAccountDto.kt
@@ -1,11 +1,12 @@
-package pt.up.fe.ni.website.backend.dto.entity
+package pt.up.fe.ni.website.backend.dto.entity.account
 
 import java.util.Date
+import pt.up.fe.ni.website.backend.dto.entity.CustomWebsiteDto
+import pt.up.fe.ni.website.backend.dto.entity.EntityDto
 import pt.up.fe.ni.website.backend.model.Account
 
-class AccountDto(
+class UpdateAccountDto(
     val email: String,
-    val password: String,
     val name: String,
     val bio: String?,
     val birthDate: Date?,
